frontend/requirements.txt
Flask==3.0.0
boto3==1.34.0
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pandas==2.1.4
pyarrow==14.0.0
